requirements.txt
langchain
chromadb
openai
tqdm
sentence-transformers
transformers
accelerate
python-dotenv
PyPDF2
streamlit
gradio
rouge-score
bert-score
trulens-eval
langgraph
langsmith
langchain-core
langchain-community
langchain-openai 
langchain-text-splitters
pdfplumber